+++ internal/services/media/media_content_key_policy_resource.go
@@ -460,7 +460,7 @@ func resourceMediaContentKeyPolicyRead(d *pluginsdk.ResourceData, meta interface
 	}
 
 	d.Set("name", id.ContentKeyPolicyName)
-	d.Set("media_services_account_name", id.AccountName)
+	d.Set("media_services_account_name", id.MediaServiceName)
 	d.Set("resource_group_name", id.ResourceGroupName)
 
 	if model := resp.Model; model != nil {